.history/inference/requirements_20250921151623.txt
ultralytics==8.3.0
fastapi
uvicorn[standard]
python-multipart
transformers
accelerate
sentencepiece
safetensors
Pillow
qrcode[pil]
supabase==2.6.0
imagehash
opencv-python-headless